fix: Correct bounds for looping variable `n`
The looping bounds for `n` are based on the assumption that `a <= b`.
However, this is not an actual constraint of Euclid's formula. This
means that the algorithm currently cannot generate all pythagorean
triples.

Change the lower bound of `n` to be based on `b >= min_factor` instead.
To rule out duplicate triples, also add the additional constraint that
`m` and `n` need to have different oddities.